Separate adopt command from offer to use cooldown
